[review] Move DepositPayment formatting to client
