+++ tests/golem/ethereum/test_transactionsystem.py
@@ -502,35 +502,3 @@ class FaucetTest(TestCase):
         assert tETH_faucet_donate(addr) is True
         assert get.call_count == 1
         assert addr in get.call_args[0][0]
-
-
-class DepositPaymentsListTest(testutils.DatabaseFixture):
-    def test_empty(self):
-        self.assertEqual(TransactionSystem.get_deposit_payments_list(), [])
-
-    def test_one(self):
-        tx_hash = \
-            '0x5e9880b3e9349b609917014690c7a0afcdec6dbbfbef3812b27b60d246ca10ae'
-        value = 31337
-        ts = 1514761200.0
-        dt = datetime.datetime.fromtimestamp(ts)
-        model.DepositPayment.create(
-            value=value,
-            tx=tx_hash,
-            created_date=dt,
-            modified_date=dt,
-        )
-        expected = [
-            {
-                'created': ts,
-                'modified': ts,
-                'fee': None,
-                'status': 'awaiting',
-                'transaction': tx_hash,
-                'value': str(value),
-            },
-        ]
-        self.assertEqual(
-            expected,
-            TransactionSystem.get_deposit_payments_list(),
-        )
